+++ packages/postgrest-core/src/index.ts
@@ -13,6 +13,7 @@ export * from './lib/get';
 export * from './lib/set-filter-value';
 export * from './lib/parse-value';
 export * from './lib/parse-order-by-key';
+export * from './lib/parse-order-by';
 export * from './lib/find-filters';
 
 export * from './cursor-pagination-fetcher';
